Moving csv_creator func to csv.py
